Refactored some files, Added uAuth
